feat(desktop): add Git settings tab
